refactor(styles): restructure styles folder to make it easier to navigate through it
