+++ galasa-ui/src/components/common/StatusIndicator.tsx
@@ -14,7 +14,7 @@ import {
   WarningFilled,
 } from '@carbon/icons-react';
 
-import styles from '@/styles/StatusIndicator.module.css';
+import styles from '@/styles/common/StatusIndicator.module.css';
 import React from 'react';
 
 interface StatusIndicatorProps {
