+++ galasa-ui/src/components/mysettings/DateTimeSettings.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 'use client';
-import styles from '@/styles/DateTimeSettings.module.css';
+import styles from '@/styles/mysettings/DateTimeSettings.module.css';
 import { useTranslations } from 'next-intl';
 import FormatSection from './FormatSection';
 import TimezoneSection from './TimezoneSection';
