+++ galasa-ui/src/components/profile/ProfileDetails.tsx
@@ -8,7 +8,7 @@
 import ErrorPage from '@/app/error/page';
 import { useDateTimeFormat } from '@/contexts/DateTimeFormatContext';
 import { UserData } from '@/generated/galasaapi';
-import styles from '@/styles/MyProfile.module.css';
+import styles from '@/styles/profile/MyProfile.module.css';
 import { ProfileDetailsProps } from '@/utils/interfaces';
 import { Loading } from '@carbon/react';
 import { useTranslations } from 'next-intl';
