+++ galasa-ui/src/components/test-runs/TestRunsTabs.tsx
@@ -5,7 +5,7 @@
  */
 'use client';
 import { Tabs, Tab, TabList, TabPanels, TabPanel } from '@carbon/react';
-import styles from '@/styles/TestRunsPage.module.css';
+import styles from '@/styles/test-runs/TestRunsPage.module.css';
 import TimeframeContent, { calculateSynchronizedState } from './timeframe/TimeFrameContent';
 import TestRunsTable from './results/TestRunsTable';
 import SearchCriteriaContent from './search-criteria/SearchCriteriaContent';
