+++ galasa-ui/src/components/test-runs/graph/TestRunsGraph.tsx
@@ -10,10 +10,10 @@ import { ScatterChart } from '@carbon/charts-react';
 import { ScaleTypes, TimeIntervalNames } from '@carbon/charts';
 import { InlineNotification } from '@carbon/react';
 import { SkeletonText } from '@carbon/react';
-import { useMemo, useRef, useEffect, useState } from 'react';
+import { useMemo, useRef, useEffect } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useTranslations } from 'next-intl';
-import styles from '@/styles/TestRunsGraph.module.css';
+import styles from '@/styles/test-runs/graph/TestRunsGraph.module.css';
 import { runStructure, ColumnDefinition, DataPoint } from '@/utils/interfaces';
 import { COLORS, MAX_RECORDS } from '@/utils/constants/common';
 import { TEST_RUNS } from '@/utils/constants/breadcrumb';
