+++ galasa-ui/src/components/test-runs/test-run-details/InlineText.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 import React from 'react';
-import styles from '@/styles/OverviewTab.module.css';
+import styles from '@/styles/test-runs/test-run-details/InlineText.module.css';
 
 const InlineText = ({ title, value }: { title: string; value: string }) => {
   return (
